Implemented add new playlist
I have created an xml layout for a landing page. After a successfull login, the user will land on this page, and he can choose, if he create a new playlist, or go to the music library.

Also, i have created the background logic of create a new playlist, BUT, we have to improve it, to put a record to the user_playlist connection table, with the current playlistID and the current userID